Split RPC blueprint
