+++ test/rpc/test_rpc_blocks.py
@@ -5,8 +5,8 @@ from test.shared import GENESIS_BLOCK_NUMBER, INCORRECT_GENESIS_BLOCK_HASH
 
 import pytest
 
-from starknet_devnet.blueprints.rpc import BlockNumberDict, BlockHashDict, rpc_txn_type
 from starknet_devnet.general_config import DEFAULT_GENERAL_CONFIG
+from starknet_devnet.blueprints.rpc_utils import BlockNumberDict, BlockHashDict, rpc_txn_type
 
 from .rpc_utils import rpc_call, pad_zero, gateway_call
