+++ test/rpc/test_rpc_class.py
@@ -3,7 +3,7 @@ Tests RPC contract class
 """
 from starkware.starknet.services.api.gateway.transaction_utils import decompress_program
 
-from starknet_devnet.blueprints.rpc import BlockId
+from starknet_devnet.blueprints.rpc_utils import BlockId
 from .rpc_utils import rpc_call
